Move DirectJobBundleFactory to fn-execution
Not required, maybe convenient.
