+++ runners/java-fn-execution/src/main/java/org/apache/beam/runners/fnexecution/control/SingleEnvironmentInstanceJobBundleFactory.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.beam.runners.direct.portable;
+package org.apache.beam.runners.fnexecution.control;
 
 import com.google.common.collect.Iterables;
 import java.io.IOException;
@@ -28,14 +28,7 @@ import org.apache.beam.model.fnexecution.v1.BeamFnApi.Target;
 import org.apache.beam.model.pipeline.v1.RunnerApi.Environment;
 import org.apache.beam.runners.core.construction.graph.ExecutableStage;
 import org.apache.beam.runners.fnexecution.GrpcFnServer;
-import org.apache.beam.runners.fnexecution.control.JobBundleFactory;
-import org.apache.beam.runners.fnexecution.control.OutputReceiverFactory;
-import org.apache.beam.runners.fnexecution.control.ProcessBundleDescriptors;
 import org.apache.beam.runners.fnexecution.control.ProcessBundleDescriptors.ExecutableProcessBundleDescriptor;
-import org.apache.beam.runners.fnexecution.control.RemoteBundle;
-import org.apache.beam.runners.fnexecution.control.RemoteOutputReceiver;
-import org.apache.beam.runners.fnexecution.control.SdkHarnessClient;
-import org.apache.beam.runners.fnexecution.control.StageBundleFactory;
 import org.apache.beam.runners.fnexecution.data.GrpcDataService;
 import org.apache.beam.runners.fnexecution.data.RemoteInputDestination;
 import org.apache.beam.runners.fnexecution.environment.EnvironmentFactory;
@@ -48,13 +41,21 @@ import org.apache.beam.sdk.fn.IdGenerators;
 import org.apache.beam.sdk.fn.data.FnDataReceiver;
 import org.apache.beam.sdk.util.WindowedValue;
 
-/** A {@link JobBundleFactory} for the ReferenceRunner. */
-class DirectJobBundleFactory implements JobBundleFactory {
+/**
+ * A {@link JobBundleFactory} which can manage a single instance of an {@link Environment}.
+ *
+ * @deprecated replace with a {@link DockerJobBundleFactory} when appropriate if the {@link
+ *     EnvironmentFactory} is a {@link
+ *     org.apache.beam.runners.fnexecution.environment.DockerEnvironmentFactory}, or create an
+ *     {@code InProcessJobBundleFactory} and inline the creation of the environment if appropriate.
+ */
+@Deprecated
+public class SingleEnvironmentInstanceJobBundleFactory implements JobBundleFactory {
   public static JobBundleFactory create(
       EnvironmentFactory environmentFactory,
       GrpcFnServer<GrpcDataService> data,
       GrpcFnServer<GrpcStateService> state) {
-    return new DirectJobBundleFactory(environmentFactory, data, state);
+    return new SingleEnvironmentInstanceJobBundleFactory(environmentFactory, data, state);
   }
 
   private final EnvironmentFactory environmentFactory;
@@ -69,7 +70,7 @@ class DirectJobBundleFactory implements JobBundleFactory {
 
   private final IdGenerator idGenerator = IdGenerators.incrementingLongs();
 
-  private DirectJobBundleFactory(
+  private SingleEnvironmentInstanceJobBundleFactory(
       EnvironmentFactory environmentFactory,
       GrpcFnServer<GrpcDataService> dataService,
       GrpcFnServer<GrpcStateService> stateService) {
@@ -114,7 +115,7 @@ class DirectJobBundleFactory implements JobBundleFactory {
             descriptor.getProcessBundleDescriptor(),
             (RemoteInputDestination<WindowedValue<T>>) (RemoteInputDestination) destination,
             stateService.getService());
-    return new DirectStageBundleFactory<>(descriptor, bundleProcessor);
+    return new BundleProcessorStageBundleFactory<>(descriptor, bundleProcessor);
   }
 
   @Override
@@ -136,11 +137,11 @@ class DirectJobBundleFactory implements JobBundleFactory {
     }
   }
 
-  private static class DirectStageBundleFactory<T> implements StageBundleFactory<T> {
+  private static class BundleProcessorStageBundleFactory<T> implements StageBundleFactory<T> {
     private final ExecutableProcessBundleDescriptor descriptor;
     private final SdkHarnessClient.BundleProcessor<T> processor;
 
-    private DirectStageBundleFactory(
+    private BundleProcessorStageBundleFactory(
         ExecutableProcessBundleDescriptor descriptor,
         SdkHarnessClient.BundleProcessor<T> processor) {
       this.descriptor = descriptor;
